+++ core/controllers/feedback.py
@@ -17,7 +17,6 @@
 from core.controllers import base
 from core.domain import acl_decorators
 from core.domain import exp_services
-from core.domain import feedback_domain
 from core.domain import feedback_services
 from core.platform import models
 import feconf
@@ -234,8 +233,7 @@ class FeedbackThreadViewEventHandler(base.BaseHandler):
 
     @acl_decorators.can_comment_on_feedback_thread
     def post(self, thread_id):
-        exploration_id = (
-            feedback_domain.FeedbackThread.get_exp_id_from_thread_id(thread_id))
+        exploration_id = feedback_services.get_exp_id_from_thread_id(thread_id)
         transaction_services.run_in_transaction(
             feedback_services.clear_feedback_message_references, self.user_id,
             exploration_id, thread_id)
