+++ core/storage/feedback/gae_models.py
@@ -18,7 +18,6 @@
 
 import datetime
 
-from core.domain import feedback_domain
 from core.platform import models
 import feconf
 import utils
@@ -494,11 +493,12 @@ class SuggestionModel(base_models.BaseModel):
 
     @classmethod
     def create(
-            cls, thread_id, author_id, exploration_version,
+            cls, exploration_id, thread_id, author_id, exploration_version,
             state_name, description, suggestion_html):
         """Creates a new SuggestionModel entry.
 
         Args:
+            exploration_id: str. ID of the exploration.
             thread_id: str. ID of the corresponding thread.
             author_id: str. ID of the user who submitted the suggestion.
             exploration_version: int. exploration version for which the
@@ -517,9 +517,6 @@ class SuggestionModel(base_models.BaseModel):
                             'thread id: %s' % instance_id)
         state_content = cls._convert_suggestion_html_to_legacy_state_content(
             suggestion_html)
-        exploration_id = (
-            feedback_domain.FeedbackThread.get_exp_id_from_thread_id(
-                instance_id))
         cls(id=instance_id, author_id=author_id,
             exploration_id=exploration_id,
             exploration_version=exploration_version,
